run test and validate
