+++ tests/aws/services/apigateway/test_apigateway_basic.validation.json
@@ -8,12 +8,6 @@
   "tests/aws/services/apigateway/test_apigateway_basic.py::TestAPIGateway::test_apigateway_with_custom_authorization_method": {
     "last_validated_date": "2024-04-12T22:31:50+00:00"
   },
-  "tests/aws/services/apigateway/test_apigateway_basic.py::TestAPIGateway::test_apigateway_with_step_function_integration[DeleteStateMachine]": {
-    "last_validated_date": "2023-09-07T20:40:38+00:00"
-  },
-  "tests/aws/services/apigateway/test_apigateway_basic.py::TestAPIGateway::test_apigateway_with_step_function_integration[StartExecution]": {
-    "last_validated_date": "2023-09-07T20:40:26+00:00"
-  },
   "tests/aws/services/apigateway/test_apigateway_basic.py::TestAPIGateway::test_apigw_test_invoke_method_api": {
     "last_validated_date": "2024-02-04T18:48:24+00:00"
   },
